refactor(angles): move functions to own file
